src/app/requirements.txt
PyJWT==2.10.1
boto3==1.35.0